requirements.txt
fastapi==0.110.2
keras==2.15.0
keras==3.2.1
numpy==1.24.4
Pillow==10.3.0
tensorflow==2.16.1
